% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{add_cTEC_mTEC_signatures}
\alias{add_cTEC_mTEC_signatures}
\title{Using predefined genesets, form simple scores summarizing cTEC and mTEC transcriptomic signatures.}
\usage{
add_cTEC_mTEC_signatures(dge)
}
\arguments{
\item{dge}{Seurat object}
}
\description{
Using predefined genesets, form simple scores summarizing cTEC and mTEC transcriptomic signatures.
}
\details{
TEC = thymic epithelial cell; c = cortical; m = medullary
}
