% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_maehrlab_metadata}
\alias{add_maehrlab_metadata}
\title{Given a Seurat object, add information about our experiments.}
\usage{
add_maehrlab_metadata(dge, variable_to_add, new_name = NULL,
  NA_strings = c("NA", ""),
  maehrlab_metadata = thymusatlasdatapublic::get_metadata())
}
\description{
Given a Seurat object, add information about our experiments.
}
\details{
It looks in `maehrlab_metadata` for a column named `variable_to_add` and adds that to @data.info with the 
name `new_name` (default is `variable_to_add`).
It relies on being able to match "Sample_ID" to "orig.ident".
}
