% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{add_pseudotime_to_seurat}
\alias{add_pseudotime_to_seurat}
\title{Transfer data from a pseudotime modeling object to a Seurat object.}
\usage{
add_pseudotime_to_seurat(dge, pt_obj, pt_method = "monocle")
}
\description{
Transfer data from a pseudotime modeling object to a Seurat object.
}
\details{
This function takes Seurat object and a finished object from a pseudotime analysis package
It transfers info from the latter to the former, guaranteeing that the Seurat object
will have complete metadata fields of the following names:
`pseudotime`, `branch`, `branch_viz_1`, `branch_viz_2`
}
