% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{add_quadrants}
\alias{add_quadrants}
\title{Split a scatterplot (or similar) into quadrants and label percentages in each quadrant.}
\usage{
add_quadrants(p, cutoffs, facet_by = NULL)
}
\arguments{
\item{cutoffs}{numeric vector of length 2. Where to delineate the quadrants.}

\item{facet_by}{optional facetting variable. Percents are calculated separately for each facet.

This is a helper for TACS, but it's exported in case it becomes useful.}

\item{dge}{Seurat object}
}
\description{
Split a scatterplot (or similar) into quadrants and label percentages in each quadrant.
}
