% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{add_rp_mt_percentage}
\alias{add_rp_mt_percentage}
\title{Quantify ribosomal and mitochondrial transcript content cell by cell.}
\usage{
add_rp_mt_percentage(dge)
}
\description{
Quantify ribosomal and mitochondrial transcript content cell by cell.
}
