% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{add_rp_percentage}
\alias{add_rp_percentage}
\title{Quantify ribosomal transcript content cell by cell.}
\usage{
add_rp_percentage(dge)
}
\description{
Quantify ribosomal transcript content cell by cell.
}
