% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_tcr}
\alias{add_tcr}
\title{Add T-cell receptor expression from a separate alignment process.}
\usage{
add_tcr(dge, metadata = thymusatlasdataprivate::get_metadata())
}
\arguments{
\item{dge}{Seurat object}

\item{metadata}{Sample sheet containing a field "Sample_ID" (to be matched with "orig.ident" from dge) 
and another called "tcr_dge_path" (to look for TCR-containing DGEs).}
}
\description{
Add T-cell receptor expression from a separate alignment process.
}
