% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{aggregate.nice}
\alias{aggregate.nice}
\title{Aggregate data.frame by a categorical variable, permissively.}
\usage{
\method{aggregate}{nice}(x, by, FUN, ...)
}
\description{
Aggregate data.frame by a categorical variable, permissively.
}
\details{
A wrapper for aggregate(). Accepts atomic "by" argument.
Guaranteed to return a dense dataframe. Also returns the aggregation levels in the rownames 
instead of adding a column for them.
}
