% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{aggregate_nice}
\alias{aggregate_nice}
\title{Aggregate data.frame by a categorical variable, permissively.}
\usage{
aggregate_nice(x, by, FUN, ...)
}
\description{
Aggregate data.frame by a categorical variable, permissively.
}
\details{
A wrapper for aggregate(). Accepts atomic "by" argument.
Guaranteed to return a matrix. Also returns the aggregation levels in the rownames 
instead of adding a column for them.
}
