% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{annot_ident_plot}
\alias{annot_ident_plot}
\title{Annotate a plot with cluster labels}
\usage{
annot_ident_plot(dge, results_path, figname, ident.use = "ident",
  height = 7, width = 8, ...)
}
\description{
Annotate a plot with cluster labels
}
