% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{are_compatible}
\alias{are_compatible}
\title{Helper function. Check if a list of markers is compatible with a given Seurat object 
so that genes and cluster assignments are present in both `marker_info` and
the Seurat object `dge`.}
\usage{
are_compatible(dge, marker_info, ident.use)
}
\description{
If `desired_cluster_order` is given, `are_compatible` checks that it is 
free of duplicates and it is a superset of the identity values occurring in other inputs.
}
