% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{call_monocle_on_seurat}
\alias{call_monocle_on_seurat}
\title{Extract data from a Seurat object and run Monocle, returning a Monocle object.}
\usage{
call_monocle_on_seurat(dge, results_path, monocle_params, earliest_day = NULL)
}
\description{
Extract data from a Seurat object and run Monocle, returning a Monocle object.
}
