% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_all_scRNA}
\alias{check_all_scRNA}
\title{Global quality control: check for Y genes, Xist and basic summary stats.}
\usage{
check_all_scRNA(results_path)
}
\description{
Global quality control: check for Y genes, Xist and basic summary stats.
}
