% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_xist}
\alias{check_xist}
\title{Check for female-specific and male-specific transcripts.}
\usage{
check_xist(raw_dge, rep_name, results_path)
}
\description{
Check for female-specific and male-specific transcripts.
}
\details{
Accepts a numeric matrix of raw molecule counts with genes in the rownames.
}
