% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_xist_all}
\alias{check_xist_all}
\title{Screen for female embryos using simple stats.}
\usage{
check_xist_all(results_path, samples = NULL, metadata = get_metadata())
}
\description{
Screen for female embryos using simple stats.
}
