% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classify_mlr}
\alias{classify_mlr}
\title{Apply a machine learning classifier (from `train_save_classifier`) to new data}
\usage{
classify_mlr(dge, mlr_mod)
}
\arguments{
\item{dge}{a Seurat object}

\item{mlr_mod}{a glmnet multiclass logistic regression model.
You can feed the output of `train_save_classifier` into the `mlr_mod` argument.}
}
\description{
Apply a machine learning classifier (from `train_save_classifier`) to new data
}
