% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{cluster_wrapper}
\alias{cluster_wrapper}
\title{A wrapper for Seurat's clustering functions.}
\usage{
cluster_wrapper(dge, results_path, test_mode, pc.use = NULL,
  method = c("DBSCAN"), granularities_as_string = "6", merge_small = F)
}
\arguments{
\item{method}{is either "DBSCAN" or "SNN".}

\item{granularities_as_string}{should be numbers separated by commas and whitespace.
The number of clusterings performed is the length of (the split and cleaned version of) `granularities_as_string`.
The granularity number is used as the neighborhood radius in DBSCAN or the "resolution" in SNN.}
}
\description{
A wrapper for Seurat's clustering functions.
}
\details{
For more information on density-based clustering (DBSCAN), look at the KDD-96 paper by Ester, Kriegel, Sander, and Xu.
"A density-based algorithm for discovering clusters in large spatial databases with noise."
For more information on the SNN-based clustering, look at (the parameter gamma in)
"A unified approach to mapping and clustering of bibliometric networks",
Ludo Waltman, Nees Jan van Eck, and Ed C.M. Noyons
}
