% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{ClusterBag}
\alias{ClusterBag}
\title{Apply a bagged clustering procedure to a Seurat object using PCA as input.}
\usage{
ClusterBag(dge, k, num_pc = NULL, feature_names = NULL, my_seed = 100,
  B = 100, ...)
}
\arguments{
\item{dge}{A Seurat object.}

\item{k}{Number of clusters. (Some may end up empty.)}

\item{num_pc}{Number of PC's to use from dge. Specify this or feature_names, not both. Default 25.}

\item{feature_names}{Fetched via FetchData as input. Specify this or num_pc, not both. Default unused.}

\item{my_seed}{RNG seed.}

\item{B}{Number of bootstrap replicates.}

\item{...}{Additional args passed to clue::cl_bag.}
}
\description{
Apply a bagged clustering procedure to a Seurat object using PCA as input.
}
