% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{ClusterBagStability}
\alias{ClusterBagStability}
\title{Wrapper for ClusterBag. Measures stability over a range of model sizes.}
\usage{
ClusterBagStability(dge, k_max = 15, ...)
}
\description{
Wrapper for ClusterBag. Measures stability over a range of model sizes.
}
