% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_species_dge}
\alias{convert_species_dge}
\title{Convert a raw digital gene expression matrix from one species to another.}
\usage{
convert_species_dge(raw_dge, from = "human", to = "mouse")
}
\description{
Convert a raw digital gene expression matrix from one species to another.
}
\details{
If two genes have the same ortholog, the molecule counts get added.
If a gene has no ortholog, it is omitted.
The input must be a matrix with genes stored in rownames( raw_dge ).
}
