% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{custom.make.unique}
\alias{custom.make.unique}
\title{Consistently label replicates}
\usage{
custom.make.unique(s, ...)
}
\description{
Consistently label replicates
}
\details{
We use the convention that the sample ID, which goes into the orig.ident variable,
looks like "e13_5_<irregular>". This helps form a consistent way of labeling replicates.
}
