% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{custom_feature_plot}
\alias{custom_feature_plot}
\title{Make tSNE plots (or PCA, or Monocle; it's customizable)}
\usage{
custom_feature_plot(dge, colour = NULL, subset_id = NULL,
  axes = c("tSNE_1", "tSNE_2"), alpha = 1, cols.use = blue_gray_red,
  use_rank = F, overplot_adjust = F, ...)
}
\description{
I needed a little bit more flexibility than Seurat was giving me with the feature plots.
Note: this used to plot the ranks of the data, but now it doesn't unless you specify `use_rank = T`.
If you want cols.use to work for categorical variables, then it should be named with the variable's levels.
For a blank plot (default), set `colour = NULL`.
}
