% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{dendrogram_merge_points}
\alias{dendrogram_merge_points}
\title{Plot a dendrogram and also cut it to merge input into `num_desired` groups.}
\usage{
dendrogram_merge_points(X, num_desired, results_path, FUN = function(x)
  hclust(dist(x), method = "ward.D2"), REORDER_FUN = function(hc)
  as.hclust(stats::reorder(as.dendrogram(hc), 1:nrow(X))),
  PLOT_FUN = plot_dendro_with_rect, CUT_FUN = stats::cutree,
  main = "Dendrogram", return_hc = F, ...)
}
\description{
Plot a dendrogram and also cut it to merge input into `num_desired` groups.
}
