% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{deseuratify_raw_data}
\alias{deseuratify_raw_data}
\title{Extract mostly-raw (medium rare?) data from a Seurat object.}
\usage{
deseuratify_raw_data(seurat_dge, retrieve_anew = F)
}
\description{
Seurat preserves the raw data exactly. Sometimes that's not ideal.
This function helps you get rawish data that have undergone the same QC filters as the Seurat scale.data,
so some cells and genes are filtered out. 
But, the numbers are UMI counts, integers, not logged or with any of that normalization BS.
This function guarantees output with `colnames(output) == dge@cell.names`.
}
\details{
Because the `@raw.data` slot was filled in wrong in some of my Seurat objects,
this can use `load_thymus_profiling_data` to get the raw data.
To toggle this behavior, set `retrieve_anew = {T,F}`.
}
