% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{DESummaryFast}
\alias{DESummaryFast}
\title{Quickly compute summary statistics for all genes.}
\usage{
DESummaryFast(dge, ident.use = "ident", aggregator = mean,
  genes_per_cluster = NULL)
}
\arguments{
\item{dge}{Seurat object}

\item{ident.use}{Any factor variable available via FetchData(dge).}

\item{aggregator}{Function used to aggregate within clusters. Try `prop_nz`.}

\item{genes_per_cluster}{Limit on number of genes returned per cluster.}
}
\description{
Quickly compute summary statistics for all genes.
}
