% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dir.create.nice}
\alias{dir.create.nice}
\title{Create new directories recursively and without a warning if they already exist.}
\usage{
dir.create.nice(my_dir)
}
\description{
Create new directories recursively and without a warning if they already exist.
}
