% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{do_enrichr}
\alias{do_enrichr}
\title{Programmatically feed gene-sets to Enrichr and save formatted results.}
\usage{
do_enrichr(results_path, geneset, geneset_name, desired_db = c("KEGG_2016",
  "WikiPathways_2016", "Reactome_2016", "BioCarta_2016", "Panther_2016",
  "NCI-Nature_2016", "GO_Biological_Process_2015"), N_ANNOT_PER_DB = 2)
}
\description{
Programmatically feed gene-sets to Enrichr and save formatted results.
}
