% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{expected_doublet_counts}
\alias{expected_doublet_counts}
\title{Given a named vector x with counts of various cell types, returns expected doublet quantities for each possible pairing.}
\usage{
expected_doublet_counts(x, rate = 0.05)
}
\description{
Given a named vector x with counts of various cell types, returns expected doublet quantities for each possible pairing.
}
\details{
Make sure you only feed this one replicate at a time! You can't get doublets across replicates.
Assumes a doublet rate of 5%. Your mileage (and flow rates) may vary.

Output is a named numeric vector of expected cell counts. 
For names, every combination of names(x) should be present once in the output.
Order doesn't matter, so labels get alphabetized and concatenated with '_'.
Within-cluster doublets are included.
E.g. you get BLD_END and BLD_BLD but not END_BLD.
}
