% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{explore_embeddings}
\alias{explore_embeddings}
\title{Quickly explore many parameter settings}
\usage{
explore_embeddings(dge, results_path, all_params, test_mode = F)
}
\description{
Quickly explore many parameter settings
}
