% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{ExploreIteratively}
\alias{ExploreIteratively}
\title{Explore a single-cell dataset, alternating between gene selection and clustering.}
\usage{
ExploreIteratively(dge, log_fc, eligible_genes = NULL, maxiter = 5,
  tol = 10, verbose = T, cluster_method = function(...)
  Seurat::FindClusters(..., print.output = F), num_pc = 25, ...)
}
\arguments{
\item{dge}{Seurat object.}

\item{log_fc}{Passed to RefineVariableGenes.}

\item{eligible_genes}{Restricts the set of genes used. Try setting it to `get_mouse_tfs()`.}

\item{maxiter}{Iteration limit. Issues warning if hit.}

\item{tol}{If the list changes by less than this many genes (added + removed), it stops.}

\item{cluster_method}{A user-provided function to re-estimate clusters, modeled off Seurat::FindClusters. 
It should take a Seurat object as the first arg.
It should also accept a `pc.use` input (even if you just discard it).
For your convenience, the principal components get updated before this function is called.
The TSNE embedding doesn't, so don't use DBClustDimension alone.}

\item{num_pc}{Used in recomputing the number of PCs, and `1:num_pc` is passed to `cluster_method` as `pc.use`.}

\item{...}{Extra parameters passed to `cluster_method`}
}
\value{
List with names "dge" (the updated seurat object) and "iteration_stats" (numbers of genes present, added, and removed at each iteration.)
}
\description{
Explore a single-cell dataset, alternating between gene selection and clustering.
}
