% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{export_from_seurat}
\alias{export_from_seurat}
\title{Export raw and normalized data from a Seurat object.}
\usage{
export_from_seurat(dge, results_path, name)
}
\description{
Export raw and normalized data from a Seurat object.
}
