% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_plotting.R
\name{facet_plot_gene_clusters}
\alias{facet_plot_gene_clusters}
\title{Draw overlaid line plots of gene clusters from output of `smooth_and_cluster_genes`.}
\usage{
facet_plot_gene_clusters(dge, results_path, cluster_mod, kmeans_features,
  abcissae_kmeans, facet_ncol = NULL)
}
\arguments{
\item{facet_ncol}{For resulting facet plots, number of columns.}
}
\description{
Draw overlaid line plots of gene clusters from output of `smooth_and_cluster_genes`.
}
