% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{FetchDataZeroPad}
\alias{FetchDataZeroPad}
\title{FetchData but with zeroes for unavailable genes}
\usage{
FetchDataZeroPad(dge, vars.all, warn = T, ...)
}
\arguments{
\item{dge}{Seurat object}

\item{vars.all}{List of all variables to fetch. Missing entries are ignored.}

\item{...}{Other arguments to pass to FetchData}
}
\description{
FetchData but with zeroes for unavailable genes
}
\details{
This function is stupid: if you ask for "PC1" and it's not available,
it will think you're asking for a non-expressed gene, so it will return zeroes.
}
