% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{FillNA}
\alias{FillNA}
\title{Remove missing values from the metadata, issuing a warning if changes are made.}
\usage{
FillNA(object, filler = "NA")
}
\arguments{
\item{object}{Seurat object}
}
\description{
Remove missing values from the metadata, issuing a warning if changes are made.
}
