% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{FindMarkersFlex}
\alias{FindMarkersFlex}
\title{Test for markers flexibly from a Seurat object.}
\usage{
FindMarkersFlex(object, ident.use, order_by = "avg_diff", ...)
}
\arguments{
\item{object}{Seurat object.}

\item{ident.use}{Variables to fetch for use in `predicate`.}

\item{...}{Passed to Seurat::FindMarkers.}
}
\description{
Test for markers flexibly from a Seurat object.
}
\details{
Calls Seurat::FindMarkers after replacing object@ident with 
`as.character(FetchData( object, ident.use )[[1]])`.
}
