% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{fix_cluster_order}
\alias{fix_cluster_order}
\title{Check if a list of cell types is good to be used to order genes and cells in a heatmap.}
\usage{
fix_cluster_order(dge, marker_info, ident.use, desired_cluster_order = NULL)
}
\description{
Check if a list of cell types is good to be used to order genes and cells in a heatmap.
}
\details{
This means:
- no duplicates
- up to order, should be equal to union of table cluster labels and dge cluster labels
- no missing values
}
