% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{get_classifier_coefs}
\alias{get_classifier_coefs}
\title{Extract coefficients from a glmnet multiclass logistic regression model.}
\usage{
get_classifier_coefs(mlr_mod)
}
\description{
Extract coefficients from a glmnet multiclass logistic regression model.
}
