% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_macosko_cc_genes}
\alias{get_macosko_cc_genes}
\title{Return cell-cycle genes from original Drop-seq paper by Macosko et al. (2015)}
\usage{
get_macosko_cc_genes(case = "Capital")
}
\description{
Return cell-cycle genes from original Drop-seq paper by Macosko et al. (2015)
}
