% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from the thymusatlasdataprivate or thymusatlasdatapublic packages.}
\usage{
get_metadata(..., verbose = F)
}
\description{
If you have both installed, thymusatlasdataprivate is preferred.
}
