% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ortholog}
\alias{get_ortholog}
\title{Return the ortholog of a given gene or NA (if no match).
Human and mouse only.}
\usage{
get_ortholog(gene, from = "human", to = "mouse")
}
\description{
Return the ortholog of a given gene or NA (if no match).
Human and mouse only.
}
