% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ortholog_nonvec}
\alias{get_ortholog_nonvec}
\title{Return the ortholog of a given gene or NA (if no match). Not vectorized. Use get_ortholog instead.}
\usage{
get_ortholog_nonvec(gene, from = "human", to = "mouse")
}
\description{
Return the ortholog of a given gene or NA (if no match). Not vectorized. Use get_ortholog instead.
}
