% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ramilowski}
\alias{get_ramilowski}
\title{Return human receptor-ligand pairs from Ramilowski et al 2015.}
\usage{
get_ramilowski()
}
\description{
Return human receptor-ligand pairs from Ramilowski et al 2015.
}
