% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{get_similar_genes}
\alias{get_similar_genes}
\title{Find genes with expression patterns similar to the genes you've specified.}
\usage{
get_similar_genes(dge, markers, n, anticorr = F)
}
\arguments{
\item{`dge`}{: a Seurat object with field `@scale.data` filled in.}

\item{`markers`:}{a character vector; giving gene names.}

\item{`n`:}{integer; number of results to return.}

\item{`anticorr`}{: allow negatively correlated genes; defaults to `FALSE`.
Given a Seurat object and a list of gene names, this function returns genes 
that are strongly correlated with those markers.}
}
\value{
character vector.
}
\description{
Find genes with expression patterns similar to the genes you've specified.
}
