% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{harmonize_species}
\alias{harmonize_species}
\title{Given a gene list and a seurat object, return a gene list compatible with 'species' in the seurat object.}
\usage{
harmonize_species(gene_list, dge)
}
\description{
Given a gene list and a seurat object, return a gene list compatible with 'species' in the seurat object.
}
