% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{is_mt}
\alias{is_mt}
\title{Decide if genes are mitochondrial. Currently uses a crude "grep mt".}
\usage{
is_mt(genes)
}
\description{
Decide if genes are mitochondrial. Currently uses a crude "grep mt".
}
