% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{is_rp}
\alias{is_rp}
\title{Decide if a gene is a ribosomal subunit or a pre-rRNA transcript.}
\usage{
is_rp(genes)
}
\description{
Decide if a gene is a ribosomal subunit or a pre-rRNA transcript.
}
