% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{KMeansGapStat}
\alias{KMeansGapStat}
\title{Imitate Seurat::FindClusters but using k-means with the gap statistic.}
\usage{
KMeansGapStat(dge, nclust = NULL, pc.use = 1:25, iter.max = 20)
}
\arguments{
\item{dge}{Seurat object}

\item{nclust}{NULL by default, so chosen via gap statistic.}

\item{pc.use}{}
}
\description{
Imitate Seurat::FindClusters but using k-means with the gap statistic.
}
