% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{master_pt}
\alias{master_pt}
\title{Run Monocle (or simpler PCA-based pseudotime) on a Seurat object, transferring input and output automatically to and from a CellDataSet object.}
\usage{
master_pt(dge, results_path, method = "monocle", earliest_day = NULL,
  mp = list(reset_var_genes = T, log_scale_expr_thresh = 0.1, excess_disp = 1,
  num_mature_types = NULL, reduction_method = "DDRTree"), ...)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Where to save tables and plots}

\item{earliest_day}{Numeric. Used to root the Monocle trajectory; gets compared to `FetchData(dge, "eday")[[1]]` via `==`.}

\item{mp}{Passed to call_monocle_on_seurat as monocle_params arg.}

\item{...}{Additional parameters passed to call_monocle_on_seurat.}
}
\description{
Run Monocle (or simpler PCA-based pseudotime) on a Seurat object, transferring input and output automatically to and from a CellDataSet object.
}
\details{
This function will look for interesting genes of different types and save a ton of plots.
The statistical inference is outsourced to Monocle via call_monocle_on_seurat (or Seurat via pc_as_pt).
}
