% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{my_err}
\alias{my_err}
\title{Print errors or warnings but return a predictable 'not_available'.}
\usage{
my_err(e, package_tried, err_or_warn = "Error", verbose = T)
}
\description{
Helps wrap calls to thymusatlasdataprivate and thymusatlasdatapublic in tryCatch blocks.
}
