% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nonzerify}
\alias{nonzerify}
\title{Quickly define an aggregator function that works only on positive (or nonzero) elements.}
\usage{
nonzerify(FUN, default = 0, allow_negative = F)
}
\arguments{
\item{FUN}{function that accepts a numeric vector.}

\item{default}{If nonzerify(FUN, default) is applied to a list of ineligible elements, it returns this.}

\item{allow_negative}{If F (default), returned function only acts on positive numbers. Otherwise, all nonzeroes are allowed in.}
}
\description{
Quickly define an aggregator function that works only on positive (or nonzero) elements.
}
