% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{optimize_cell_order}
\alias{optimize_cell_order}
\title{Returns a new ordering of the cells (a list of cell names).
Cells will be ordered first by cluster and then (within clusters) by the sum of 
the expression levels of that cluster's markers.}
\usage{
optimize_cell_order(dge, marker_info, ident.use, desired_cluster_order)
}
\arguments{
\item{dge}{should be a Seurat object.}

\item{marker_info}{dataframe containing variables `gene` and `cluster`.}

\item{ident.use}{tells you what variable to pull from the Seurat object.
The first three args should be compatible according to `are_compatible()`.
Also, `desired_cluster_order` should be a superset of `Seurat::FetchData(dge, vars.all = ident.use)[[1]]`.}
}
\description{
Returns a new ordering of the cells (a list of cell names).
Cells will be ordered first by cluster and then (within clusters) by the sum of 
the expression levels of that cluster's markers.
}
