% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{outlier_labeled_scatterplot}
\alias{outlier_labeled_scatterplot}
\title{Return a scatterplot, labeling outliers.}
\usage{
outlier_labeled_scatterplot(data, main = "", prop_label = 0.02)
}
\value{
Prints and returns a ggplot object.
}
\description{
Return a scatterplot, labeling outliers.
}
\details{
This function plots two variables, labeling the ones that are far from their neighbors.
`data` should be a dataframe with first column to go on the horizontal axis,
   second on the vertical. The third column is treated as the labels and the
   fourth (optional) is treated as the color.
`main` shows up as the figure title.
The proportion of points labeled is `prop_label`.
}
