% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{pc_as_pt}
\alias{pc_as_pt}
\title{Take a principal component (first by default) and return a list of results gained by interpreting that PC as a pseudotime axis.}
\usage{
pc_as_pt(dge, pc.use = 1, orient_var = "eday")
}
\arguments{
\item{dge:}{Seurat object to be used. Should have variable genes and PCA fields already filled in. 
Jackstraw and t-SNE are optional but preferred.}

\item{orient_var:}{pseudotime is flipped to correlate positively with this. Anything available 
from `Seurat::FetchData(dge, orient_var)` is fair game... as long as it's numeric.}

\item{pc.use:}{which principal component to take}
}
\description{
Take a principal component (first by default) and return a list of results gained by interpreting that PC as a pseudotime axis.
}
