% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{plot_dendro_with_rect}
\alias{plot_dendro_with_rect}
\title{Plot a dendrogram, labeling merges given by `converter` with colored rectangles.}
\usage{
plot_dendro_with_rect(hc, converter, main = "Dendrogram")
}
\description{
Plot a dendrogram, labeling merges given by `converter` with colored rectangles.
}
\details{
If converter is c("a", "a", "b", "c"), you get edges labeled as a 
four-colour muted rainbow next to a grayscale with light, light, medium, and dark.
This is to convey that the first two tips have been somehow grouped.
}
