% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{RefineVariableGenes}
\alias{RefineVariableGenes}
\title{Update the `@var.genes` slot of a Seurat object using cluster markers.}
\usage{
RefineVariableGenes(dge, log_fc = 1, eligible_genes = NULL)
}
\arguments{
\item{dge}{Seurat object}

\item{log_fc}{For each gene, we compute the difference between the highest and lowest cluster mean. 
If it exceeds this, the gene is included in the active set.}

\item{eligible_genes}{dge@var.genes get intersected with this list before return.}
}
\value{
A Seurat object.
}
\description{
Update the `@var.genes` slot of a Seurat object using cluster markers.
}
