% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_species_specific_genes}
\alias{remove_species_specific_genes}
\title{Remove species-specific genes from a Seurat object.}
\usage{
remove_species_specific_genes(dge, results_path, threshold = 0.003,
  diff_thresh = 0.5)
}
\description{
Remove species-specific genes from a Seurat object.
}
\details{
The input is a Seurat object with a metadata field "species".
Removes genes that appear in less than 3 out of 1000 cells of any species
Also removes genes where the proportion in one species is higher by 0.5 than the other.
}
