% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{remove_TEC_doublets}
\alias{remove_TEC_doublets}
\title{Remove intra-TEC doublets by finding cells with both a high cTEC signature and a high mTEC signature.}
\usage{
remove_TEC_doublets(dge, results_path, mTEC_thresh = NULL,
  cTEC_thresh = NULL, pure_fraction_mTEC = 0.02, pure_fraction_cTEC = 0.2,
  reject_rate = 0.05)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Plots are saved here.}

\item{mTEC_thresh, }{@param cTEC_thresh Cells are excluded if they exceed both thresholds.
If these are NULL, they are selected automatically.}
}
\value{
A list with names "dge", "mTEC_thresh", and "cTEC_thresh."
}
\description{
Remove intra-TEC doublets by finding cells with both a high cTEC signature and a high mTEC signature.
}
\details{
For threshold selection, we isolate some nearly-pure cTECs by ranking with all cells on 
cTEC signatures and retaining the top pure_fraction_cTEC (default 20%).
Among those cells, we take the (1 - reject_rate) quantile (default 98th percentile), as the
rejection threshold. 
The same process is used for mTEC threshold selection, but pure_fraction_mTEC defaults to just 2%.
}
