% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{replace_with_int_rank}
\alias{replace_with_int_rank}
\title{Convert a, b, b, a, c, a to 1, 2, 2, 1, 3, 1. Works with any strings alphabetically.}
\usage{
replace_with_int_rank(x)
}
\description{
Convert a, b, b, a, c, a to 1, 2, 2, 1, 3, 1. Works with any strings alphabetically.
}
