% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{save_depth_stats}
\alias{save_depth_stats}
\title{Assemble very basic summary stats: total UMIs, genes, and cells.}
\usage{
save_depth_stats(results_path, samples = NULL, metadata = get_metadata())
}
\description{
Assemble very basic summary stats: total UMIs, genes, and cells.
}
