% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_feature_plots}
\alias{save_feature_plots}
\title{Save plots en masse.}
\usage{
save_feature_plots(dge, results_path, top_genes = NULL, by_cluster = NULL,
  gene_list = NULL, gene_list_name = NULL, axes = NULL,
  axes_description = NULL, do_time_series = F, alpha = 1, ...)
}
\arguments{
\item{dge}{Seurat object with available t-SNE coords (or whatever's in `axes`)}

\item{alpha}{Transparency of points}

\item{...}{Additional parameters are passed to `custom_feature_plot` or `time_series`}

\item{results_path:}{where to save the resulting plots}

\item{top_genes:}{deprecated; do not use}

\item{by_cluster:}{deprecated; do not use}

\item{gene_list:}{character vector consisting of gene names}

\item{gene_list_name:}{used in file paths so that you can call this function again with different `gene_list_name`}

\item{axes:}{any pair of numeric variables retrievable via FetchData. Defaults to `c("tSNE_1", "tSNE_2")`.}

\item{axes_description:}{used in file paths so that you can call this function again with different `axes_description` but the same `results_path` and it won't overwrite.}

\item{time_series:}{Uses `time_series` internally instead of `custom_feature_plot`. Changes defaults for
`axes` and `axes_description`.}
}
\description{
Save plots en masse.
}
