% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{save_heatmap}
\alias{save_heatmap}
\title{Save a big PDF file to `<results_path>/<main>.pdf` containing a heatmap of gene expression levels.}
\usage{
save_heatmap(dge, results_path, marker_info, desired_cluster_order = NULL,
  main = "heatmap", ident.use = "ident", do_key = F, cs = blue_yellow,
  dendrogram = F, test_mode = F)
}
\arguments{
\item{dge}{a Seurat object}

\item{results_path}{should be a character such that `dir.exists( results_path )`}

\item{marker_info}{a dataframe containing variables `gene` and `cluster`.}

\item{do_key}{If `TRUE` a color legend gets added}

\item{cs}{a vector of color names}

\item{dendrogram}{If TRUE, a cell dendrogram is added. It is constrained to be compatible with `dge@ident`. The 
implementation changes considerably. This feature is underdeveloped and other inputs -- especially `desired_cluster_order` -- may not work properly.}

\item{test_mode}{If `TRUE`, use only 100 genes and 100 cells.}

\item{desired_cluster_order:}{a superset of `levels(dge@ident)`}
}
\description{
Save a big PDF file to `<results_path>/<main>.pdf` containing a heatmap of gene expression levels.
}
\details{
Each column is a cell and each row is a gene. Each gene is rescaled so that its peak expression is 1.
This facilitates comparison within genes and across cells, though it's bad for comparison across genes.
}
