% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{save_marker_table}
\alias{save_marker_table}
\title{Clean and save a df containing columns "gene", "avg_diff", "pct.1", "pct.2", "p.value" and possibly "q.value".}
\usage{
save_marker_table(X, results_path, testname, remove_rp = FALSE,
  add_to_inventory = FALSE)
}
\description{
Clean and save a df containing columns "gene", "avg_diff", "pct.1", "pct.2", "p.value" and possibly "q.value".
}
