% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{scatterplot_replicates}
\alias{scatterplot_replicates}
\title{This function checks for consistency across replicates as returned by `get_data_by_replicates`.
It plots avg expression for each gene and proportion expressing each gene.}
\usage{
scatterplot_replicates(results_path)
}
\description{
This function checks for consistency across replicates as returned by `get_data_by_replicates`.
It plots avg expression for each gene and proportion expressing each gene.
}
