% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{SeuratMerge}
\alias{SeuratMerge}
\title{Merge two Seurat objects.}
\usage{
SeuratMerge(dge1, dge2, vars.keep = intersect(names(dge1@data.info),
  names(dge2@data.info)))
}
\description{
By default, preserves any column of @data.info shared between both objects.  
You can also specify what variables to keep. They will be added to data.info in 
the output, warning and padding with zeroes if either object lacks any var in vars.keep.
}
