% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{SeuratPie}
\alias{SeuratPie}
\title{Make small-multiple pie charts.}
\usage{
SeuratPie(dge, ident.use = "cell_type", facet_by = "eday", col = NULL,
  label = F, main = "Sample makeup by day")
}
\arguments{
\item{dge}{Seurat object}

\item{ident.use}{Becomes the categories in the pie chart}

\item{facet_by}{Each small multiple contains cases at one level of this variable.}

\item{col}{Optional colorscale.}

\item{label}{Logical. If TRUE, percentages are added.}

\item{main}{Plot title.}
}
\description{
Make small-multiple pie charts.
}
