% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{smooth_and_cluster_genes}
\alias{smooth_and_cluster_genes}
\title{Characterize genes by behavior over pseudotime, returning cluster assignments and effect sizes.}
\usage{
smooth_and_cluster_genes(dge, results_path, num_periods_initial_screen = 20,
  prop_genes_keep = 0.1, abcissae_kmeans = 20, num_clusters = NULL)
}
\arguments{
\item{dge}{should be a seurat object with a field "pseudotime". 
The field `dge@data` is accessed for expression levels -- for Eric's objects, the units will be log2(1+CP10K).}

\item{results_path}{is a character vector showing where to dump the output.}

\item{num_periods_initial_screen}{Cells are partitioned into this many pseudotime periods (equal number of cells in each).
They get averaged and each gene's "effect size" is the largest minus the smallest average.}

\item{prop_genes_keep}{Genes are ranked by effect size and the top prop_genes_keep*100% are sent into smoothers and kmeans.}

\item{abcissae_kmeans}{Gene expression is fed into k-means as a series of predictions at successive time points.
The arguments says how many time points to predict and feed in (if length one) or what time points (if longer).}

\item{num_clusters}{Genes are partitioned into this many modules. If NULL (default) the value is selected via the gap statistics and their SEs using the method in the original gap statistic paper.

      Tibshirani, R., Walther, G. and Hastie, T. (2001). Estimating the number 
      of data clusters via the Gap statistic. Journal of the Royal Statistical Society B, 63, 411–423.
                                
There's one adjustment: this function will never use just one cluster. It will issue a warning and use 2.}
}
\description{
Characterize genes by behavior over pseudotime, returning cluster assignments and effect sizes.
}
\details{
This function helps explore gene dynamics over pseudotime. It goes through three main steps:

- find genes that respond strongly to pseudotime.
- smooth those genes' expression to form an overall pseudotime trend.
- cluster genes based on smoothed expression patterns that have been shifted/scaled to the unit interval.
}
