% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{sparse_axis}
\alias{sparse_axis}
\title{Set up a sparse axis with few labels for many x/y values
Not an ideal interface (sorry!): for eight observations in groups of 5 and 3, 
the `labels` input must be like c("", "", "lab1", "", "", "", "lab2", "").}
\usage{
sparse_axis(labels, side, ...)
}
\description{
Set up a sparse axis with few labels for many x/y values
Not an ideal interface (sorry!): for eight observations in groups of 5 and 3, 
the `labels` input must be like c("", "", "lab1", "", "", "", "lab2", "").
}
