% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_path}
\alias{split_path}
\title{Split out folder names and return them in reverse order with "" at the end.}
\usage{
split_path(path)
}
\description{
Split out folder names and return them in reverse order with "" at the end.
}
\details{
split_path("/Users/erickernfeld/Dropbox/2016JULY07scRNAseq/") yields:
"2016JULY07scRNAseq"      "Dropbox"            "erickernfeld"       "Users"      ""
}
