% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{strip_suffix}
\alias{strip_suffix}
\title{Remove a suffix if it is present, but if it's absent, avoid damaging the input.}
\usage{
strip_suffix(s, suffix)
}
\description{
Remove a suffix if it is present, but if it's absent, avoid damaging the input.
}
\details{
Not vectorized! Length-1 character vectors please.
}
