% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{SubsetDataFlex}
\alias{SubsetDataFlex}
\title{Subset data flexibly from a Seurat object.}
\usage{
SubsetDataFlex(dge, vars.use, predicate, preserve_raw = FALSE,
  results_path = NULL, show_on_tsne = !is.null(results_path), ...)
}
\arguments{
\item{dge}{Seurat object}

\item{vars.use}{Variables to fetch for use in `predicate`.}

\item{predicate}{String to be parsed into an R expression and evaluated as an arg to base::subset.}

\item{preserve_raw}{If TRUE, then it will not exclude any cells from the @raw.data slot.
By default, it leaves cells in @raw.data only if they satisfy the given predicate.}

\item{results_path}{Where to save the tSNE.}

\item{show_on_tsne}{If true, save a tSNE showing which cells are kept.}

\item{...}{Extra params passed to tsne_colored.}
}
\description{
Subset data flexibly from a Seurat object.
}
\details{
Calls FetchData, subsets it as a dataframe using base::subset, and 
subsets the Seurat object correspondingly (using the df rownames.)
}
