% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{train_save_classifier}
\alias{train_save_classifier}
\title{Train and save a penalized logistic regression classifier.}
\usage{
train_save_classifier(training_dge, results_path, ident.use = "cell_type",
  genes.use, do.save = F)
}
\description{
Uses Seurat::FetchData(training_dge, vars.all = ident.use ) as class labels.
Results (`glmnet` object) and training data (Seurat object) get
saved into a subdirectory of `results_path`.
}
