% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{tsne_colored}
\alias{tsne_colored}
\title{Save plots from `custom_feature_plot`.}
\usage{
tsne_colored(dge, results_path, colour = NULL, fig_name = NULL,
  axes = c("tSNE_1", "tSNE_2"), axes_description = "TSNE", alpha = 1,
  height = 7, width = 8, types = c("PDF", "PDF_no_leg", "PNG_PDF_split"),
  ...)
}
\arguments{
\item{dge}{Seurat object}

\item{results_path}{Where to save files.}

\item{colour}{A gene or type of metadata. Numeric zeroes plotted if `!is.element( colour, AvailableData( dge ) )`.}

\item{fig_name}{Figure gets named <fig_name>.pdf or <fig_name>.png or similar. If you put a name ending in ".png" or ".pdf", the extension is stripped off.}

\item{axes}{Character vector of length 2. Name of numeric variables available from `FetchData`.}

\item{axes_description}{Character. Used in file paths, so no spaces please.}

\item{alpha}{Numeric of length 1 between 0 and 1. Point transparency.}

\item{height}{Passed to ggsave.}

\item{width}{Passed to ggsave, but when you ask for a legend, it gets stretched a bit to make up for lost horizontal space.}

\item{types}{Atomic character vector; can be longer than 1 element. If contains "PDF", you get a PDF back. If "PDF_no_leg", you get a PDF with no legend. If "PNG_PDF_split", you get back the points and bare axes in a PNG, plus text-containing elements in a PDF with no points. By default, does all three. Matching is not case sensitive.}

\item{...}{Additional arguments passed to `custom_feature_plot`.}
}
\description{
Save plots from `custom_feature_plot`.
}
