% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{wheel_plot}
\alias{wheel_plot}
\title{Visualize probabilistic classification results}
\usage{
wheel_plot(dge, results_path, mlr_mod = NULL, class_labels = NULL, fig_name,
  colour_by = NULL, facet_by = NULL, style = "points",
  wheel_order = NULL, do.density = NULL)
}
\arguments{
\item{dge}{a Seurat object}

\item{results_path}{folder to place output in}

\item{mlr_mod}{a glmnet multiclass logistic regression model. Give this or `class_labels`, not both.
You can feed the output of `train_save_classifier` into the `mlr_mod` argument.}

\item{class_labels}{atomic character vector naming columns in the Seurat object that contain
class probabilities. Give this or `mlr_mod`, not both. If names(class_labels) is filled in,
then that's how the spokes get labeled.}

\item{fig_name}{filename for output.}

\item{colour_by}{Variable in Seurat object to map color to; default is none}

\item{facet_by}{Variable in Seurat object to facet resulting plots by; default is none}

\item{style}{If "points", plot each cell as a dot. 
If "density", then instead of plotting points, plot 2d density contours.
If "hexagons", do AWESOME HEXAGON BINNING YEAHHHHHHH HEXAGONS.}

\item{wheel_order}{Deprecated.}

\item{do.density}{Deprecated.}

\item{fig_name}{filename for output.}
}
\description{
Visualize probabilistic classification results
}
