% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{add_cc_score}
\alias{add_cc_score}
\title{Quantify cell-cycle activity by phase.}
\usage{
add_cc_score(dge, method = "average")
}
\description{
The function `add_cc_score` takes a Seurat object and adds cell cycle scores to the metadata (`object@data.info`). The columns are quantitative scores for each of the five cell-cycle phases.
}
