% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{add_maehrlab_metadata}
\alias{add_maehrlab_metadata}
\title{Look in PATH_TO_METADATA for extra information about our experiments.}
\usage{
add_maehrlab_metadata(dge, variable_to_add = "eday", new_name = NULL,
  NA_strings = c("NA", ""))
}
\description{
Look in PATH_TO_METADATA for extra information about our experiments.
}
\details{
It looks for a column named `variable_to_add` and adds that to @data.info with the 
name `new_name` (default is `variable_to_add`).
It relies on being able to match "Sample_ID" to "orig.ident".
}
