% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_all_scRNA}
\alias{check_all_scRNA}
\title{Screen for female embryos using simple stats.}
\usage{
check_all_scRNA(results_path, samples = NULL,
  metadata = thymusatlasdataprivate::get_metadata())
}
\description{
Screen for female embryos using simple stats.
}
