% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{ClusterBag}
\alias{ClusterBag}
\title{Apply a bagged clustering procedure to a Seurat object using PCA as input.}
\usage{
ClusterBag(dge, k, num_pc = 25, my_seed = 100)
}
\description{
Apply a bagged clustering procedure to a Seurat object using PCA as input.
}
