% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{custom_feature_plot}
\alias{custom_feature_plot}
\title{Make tSNE plots (or PCA, or Monocle; it's customizable)}
\usage{
custom_feature_plot(dge, colour = NULL, subset_id = NULL,
  axes = c("tSNE_1", "tSNE_2"), fix_coords = T, show_axes = F,
  alpha = 1, cols.use = c("khaki1", "red"), use_rank = F,
  overplot_adjust = F)
}
\arguments{
\item{dge}{Seurat object}

\item{colour}{Variable to fetch and plot. Length-1 character. "plain_grey" returns a dark grey plot.}

\item{subset_id}{Vector of subsets to include on the plot. Cells not included will still be included via
geom_blank so as to preserve the scale of the plot.}

\item{axes}{Variables to fetch for use on X and Y axes. Defaults to tSNE embedding.}

\item{fix_coords}{Force the plot to be square? Defaults to T, because a rotated or reflected version of 
the tSNE embedding should convey the exact same relationships.}

\item{alpha}{Transparency. Number between 0 and 1.}

\item{cols.use}{Colorscale. Character vector, hexadecimal or e.g. c("khaki1", "red"). 
For discrete data fetched for "colour", should be named with the variable's levels.}

\item{use_rank}{Transform expression level to ranks before plotting?}

\item{overplot_adjust}{Bin points within hexagons. Nice for Monocle embeddings.}
}
\description{
Make tSNE plots (or PCA, or Monocle; it's customizable)
}
