% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{dge_merge_list}
\alias{dge_merge_list}
\title{Merge a list of digital gene expression matrices.}
\usage{
dge_merge_list(dge_list, allow_barcode_overlap = F)
}
\arguments{
\item{dge_list}{list of matrices with genes as rows and cell barcodes as columns.}

\item{allow_barcode_overlap}{Set this to TRUE if matching barcodes in different inputs refer to the same cells, 
as in a resequencing of a library. Set this to FALSE (default) if matching barcodes would only occur by coincidence, 
as in merging of different replicates for joint analysis.}
}
\description{
Merge a list of digital gene expression matrices.
}
\details{
`dge_merge_list` converts the digital gene expression matrices to dataframes with genes as columns, merges them, then converts the result back, all without disturbing the gene labels. A gene will be included if it appears in any of the datasets. If a gene appears in one dataset but not another, zeroes will be filled in for missing expression levels.
}
