% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{faceted_tsne}
\alias{faceted_tsne}
\title{Plot a dataset's tSNE embedding faceted by experimental metadata (default replicate and eday).}
\usage{
faceted_tsne(dge, results_path, inner_factor = "orig.ident",
  outer_factor = "eday", colour = outer_factor, width = 20, height = 12)
}
\description{
Plot a dataset's tSNE embedding faceted by experimental metadata (default replicate and eday).
}
