% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R, R/utilities.R
\name{get_cTEC_mTEC_genes}
\alias{get_cTEC_mTEC_genes}
\alias{get_cTEC_mTEC_genes}
\title{Return a table of genes generated by head-to-head comparison of cTECs and mTECs on the Atlas data.}
\usage{
get_cTEC_mTEC_genes()

get_cTEC_mTEC_genes()
}
\description{
Return a table of genes generated by head-to-head comparison of cTECs and mTECs on the Atlas data.

Return a table of genes generated by head-to-head comparison of cTECs and mTECs on the Atlas data.
}
