% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{load_thymus_profiling_data}
\alias{load_thymus_profiling_data}
\title{Get digital gene expression matrices from either thymusatlasdataprivate or thymusatlasdatapublic.}
\usage{
load_thymus_profiling_data(..., verbose = F)
}
\description{
If you have both available, thymusatlasdataprivate is preferred.
}
