% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{make_heatmap_for_table}
\alias{make_heatmap_for_table}
\title{Make a heatmap with one column for each cluster in `unique( Seurat::FetchData(dge, ident.use)[[1]])` and 
one row for every gene in `genes_in_order`.}
\usage{
make_heatmap_for_table(dge, genes_in_order, desired_cluster_order = NULL,
  ident.use = "ident", labels = NULL, aggregator = mean,
  normalize = "row", norm_fun = div_by_max,
  main = "Genes aggregated by cluster")
}
\description{
If the cluster's expression values are stored in `x`, then `aggregator(x)` gets (normalized and) plotted.
Optional parameter `desired_cluster_order` gets coerced to character. Should be a permutation of 
`unique(Seurat::FetchData(dge, ident.use))`, though elements may be omitted.
}
