% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{ProjectCells}
\alias{ProjectCells}
\title{Project cells from one Seurat object onto the pseudotime axis from another Seurat object.}
\usage{
ProjectCells(dge_train, dge_test, genes.use = dge_train@var.genes,
  regressout = c("IG1.S", "S", "G2.M", "M", "M.G1"),
  new_name = "pseudotime", method = "KNN", k = 25)
}
\arguments{
\item{dge_train}{Seurat object with 'pseudotime' field included.}

\item{dge_test}{Seurat object. Cells will be assigned a pseudotime value based on `dge_train`.}

\item{genes.use}{Defaults to `dge_train@var.genes`. If you use method="PCA", use the same genes here.
If you aren't sure, use method="KNN".}

\item{regressout}{Please put whatever you used for 'latent.vars' in RegressOut when you processed 'dge_train'.
Only matters if you use method="PCA", so if you aren't sure, use method="KNN".}

\item{new_name}{Name of column to be added to `dge_test`.}

\item{method}{Determines whether to use a PCA-based approach (if "PCA") or an average over nearest neighbors (anything else). Not case sensitive.}

\item{k}{Number of nearest neighbors to use.}
}
\description{
Project cells from one Seurat object onto the pseudotime axis from another Seurat object.
}
