% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{SubsetDataFlex}
\alias{SubsetDataFlex}
\title{Subset data flexibly from a Seurat object.}
\usage{
SubsetDataFlex(dge, vars.use, predicate, zeropad = TRUE)
}
\arguments{
\item{dge}{Seurat object}

\item{vars.use}{Variables to fetch for use in `predicate`.}

\item{predicate}{String to be parsed into an R expression and evaluated as an arg to `base::subset`.}
}
\description{
Subset data flexibly from a Seurat object.
}
\details{
Calls FetchData, subsets it as a dataframe using base::subset, and 
subsets the Seurat object correspondingly (using the df rownames.)
}
