% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_extra.R
\name{TACS}
\alias{TACS}
\title{Make a FACS-like plot from a single-cell dataset.}
\usage{
TACS(dge, gene1, gene2, return_val = "plot", num_genes_add = 100,
  facet_by = NULL, cutoffs = NULL)
}
\arguments{
\item{dge}{Seurat object}

\item{gene1}{Horizontal axis on plot mimics this gene. Character.}

\item{gene2}{Vertical axis on plot mimics this gene. Character.}

\item{return_val}{If "all", returns a list with several internal calculations revealed.
If "plot", returns just a ggplot object. If "seurat", returns a Seurat object with gene scores added.}

\item{num_genes_add}{Each axis shows a simple sum of similar genes. This is how many (before removing overlap). Integer.}

\item{cutoffs}{If given, divide plot into four quadrants and annotate with percentages. Numeric vector of length 2.}
}
\description{
Make a FACS-like plot from a single-cell dataset.
}
