% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{time_series}
\alias{time_series}
\title{Return spline-smoothed expression plots over pseudotime.}
\usage{
time_series(dge, gene, colour = "eday", main = NULL, x = "pseudotime",
  col = Thanksgiving_colors, add_points = T)
}
\description{
Return spline-smoothed expression plots over pseudotime.
}
