% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_pipeline.R
\name{time_series_save}
\alias{time_series_save}
\title{Save plots from `times_series`.}
\usage{
time_series_save(dge, results_path, gene, x = "pseudotime", types = c("pdf",
  "pdf_no_leg"), width = 8, height = 6, colour = "eday", ...)
}
\arguments{
\item{types}{For an explanation of the "types" param, see ?tnse_colored .}
}
\description{
Save plots from `times_series`.
}
